+++ packages/coding-agent/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Running bash and eval commands now show elapsed wall time next to the timeout footer.
+
 ## [18.0.3] - 2026-08-23
 
 ### Added